TITAN_PROJECT_02/TITAN_V3/requirements.txt
tensorflow==2.15.0
tensorflow-tensorrt
keras==3.0.0
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
pandas-ta==0.3.14b0
scikit-learn==1.3.2
joblib==1.3.2
yfinance==0.2.33
MetaTrader5==5.0.44
flask==2.3.3
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0
tqdm==4.66.1
pyzmq==25.1.2
